+++ Makefile
@@ -32,10 +32,12 @@ SRCS		= 	input_manager/buffer_manager.c input_manager/history.c \
 				parser/parser.c parser/command.c parser/lexer.c errors.c \
 				var.c parser/destroyer.c parser/lexer_parts.c \
 				parser/dollar.c parser/make_word.c parser/error_detector.c \
-				exec/exec.c buildin/cd.c buildin/exit.c build_args.c \
+				buildin/cd.c buildin/exit.c build_args.c \
 				buildin/echo.c buildin/env.c buildin/unset.c buildin/pwd.c \
-				buildin/export.c exec/prepare_exec.c exec/redirection.c \
-				exec/redirection_utils.c
+				buildin/export.c \
+				exec/prepare_exec.c exec/redirection.c \
+				exec/redirection_utils.c exec/exec_utils.c exec/exec_utils2.c \
+				exec/execution_rules.c exec/exec.c exec/child.c
 
 MAIN		= main.c
