+++ srcs/exec/exec.h
@@ -6,7 +6,7 @@
 /*   By: tmatis <tmatis@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/04 14:12:28 by tmatis            #+#    #+#             */
-/*   Updated: 2021/05/22 13:29:47 by tmatis           ###   ########.fr       */
+/*   Updated: 2021/05/22 19:36:09 by tmatis           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,5 +25,13 @@ void	restore_in_out(int backup[2]);
 int		redir_dup_fail(char *file, int backup[2]);
 void	restore_in_out(int backup[2]);
 int		get_open_flags(int type);
+int 	execution_rules(t_command command, t_list **env_vars);
+void 	close_all_pipes(t_tube *tube_list, int size);
+void 	close_finish_tube(t_tube *tube_list, int forks_running, int fork_n);
+void 	execution_error_status(char *cmd, char *error, int status, t_list **env_vars);
+int 	handle_dup2_error(t_tube *tube_list, int forks_running, int fork_n);
+void 	sig_handler(int signal_no);
+int		child_process(t_child_data child_data);
+
 
 #endif
